+++ changelog.md
@@ -14,6 +14,7 @@ Changelog
 + Reimplement `SegmentControl` as `RayControl`
 + Rename `control.copyAnchor` to `affixAnchor`
 + Rename `control.center` to `knob`
++ Rename `segment.withStartExtended` to `withStartExtension`
 
 
 1.0.1 - 2021 August 20th